Template for testing driver endpoints.
